requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
supabase==2.4.0
SQLAlchemy==2.0.46
Werkzeug==3.1.5
gunicorn==22.0.0
